+++ custom_model_runner/CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-#### [1.4.16] - in progress
+#### [1.4.16] - 2021-02-18
 ##### Added
 - `/info` endpoint
+- progress spinner when building a docker images
+
+##### Fixed
+- unset entrypoint when running DRUM in docker
 
 #### [1.4.15] - 2021-02-11
 ##### Fixed